Added links to next and previous week, refactored, added some styles
Each schedule page now links to both the previous and next week's
schedule.

I also did some refactoring to improve function names and added a bit
of CSS to make the "Cock-a-doodle-doo!" button change color when
hovered over or clicked on.

This commit should probably have been split up into multiple commits,
but I decided not to go through the trouble of splitting it up.